Refactoring: pass from RectF to Rect.
Because float algebra is slower then integer i decide to use only int for hitbox

+++ app/src/main/java/com/mazerfaker/pewpewboom/model/Background.java
@@ -40,7 +40,7 @@ public class Background {
     }
 
 
-    private float _y;
+    private int _y;
     private Bitmap _bitmap;
     private int _bitmapHeight;
     private int _backgroundSpeed;
