+++ app/src/main/java/com/mazerfaker/pewpewboom/model/weapons/Blaster.java
@@ -15,7 +15,7 @@ public class Blaster extends BaseWeapon implements Weapon {
 
 
     @Override
-    public StandardBullet fire(float x, float y) {
+    public StandardBullet fire(int x, int y) {
         tick();
 
         if(_fireCount != 0) {
