+++ app/src/main/java/com/mazerfaker/pewpewboom/model/weapons/Weapon.java
@@ -5,5 +5,5 @@ import com.mazerfaker.pewpewboom.model.characters.Bullet;
 
 public interface Weapon {
 
-    public Bullet fire(float x, float y);
+    public Bullet fire(int x, int y);
 }
